# docs/openapi/agent-v1.yaml
openapi: 3.1.0
info:
  title: Fidence Merchant API v1
  version: 1.0.0
  description: |
    Minimal OpenAPI spec listing Fidence agent and merchant API routes under `/api/v1`.
    Authenticate with `Authorization: Bearer fid_live_...` unless noted otherwise.
servers:
  - url: https://payagent.co/api/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /agents/{agentId}:
    patch:
      tags: [Agents]
      summary: Update agent metadata or status
      operationId: patchAgent
      parameters:
        - $ref: "#/components/parameters/agentId"
      responses:
        "200":
          description: Agent updated
...
  /payment-links/{linkId}:
    get:
      tags: [Payment Links]
      summary: Get a payment link
      operationId: getPaymentLink
      parameters:
        - name: linkId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment link
...
  /pay:
    post:
      tags: [Pay]
      summary: Record an agent payment after on-chain settlement
      operationId: recordAgentPay
      responses:
        "200":
          description: Payment recorded
        "202":
          description: Approval required
...
  /webhooks/{id}:
    patch:
      tags: [Webhooks]
      summary: Update a webhook endpoint
      operationId: patchWebhook
      parameters:
        - $ref: "#/components/parameters/webhookId"
      responses:
        "200":
          description: Webhook updated
    delete:
      tags: [Webhooks]
      summary: Delete a webhook endpoint
      operationId: deleteWebhook
      parameters:
        - $ref: "#/components/parameters/webhookId"
      responses:
        "200":
          description: Webhook deleted
...
  /compliance/agents/{agentId}/policy:
    get:
      tags: [Compliance]
      summary: Get agent policy
      operationId: getAgentPolicy
      parameters:
        - $ref: "#/components/parameters/agentId"
      responses:
        "200":
          description: Policy document
    put:
      tags: [Compliance]
      summary: Create or update agent policy
      operationId: putAgentPolicy
      parameters:
        - $ref: "#/components/parameters/agentId"
      responses:
        "200":
          description: Policy saved
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: fid_live_...
  parameters:
    agentId:
      name: agentId
      in: path
      required: true
      schema:
        type: string
      description: External agent id (e.g. checkout-bot)
    webhookId:
      name: id
      in: path
      required: true
      schema:
        type: string
...